requirements.txt
amqp==2.6.1
asgiref==3.2.10
autopep8==1.5.6
billiard==3.6.3.0
boto3==1.17.15
botocore==1.20.15
celery==4.4.7
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
cloudinary==1.24.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.7
Django==2.2.16
django-allauth==0.45.0
django-appconf==1.0.4
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-extensions==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
django-hosts==4.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-multitenant==2.3.2
django-referrer-policy==1.0
django-request-logging==0.7.2
django-requestlogging==1.0.1
django-rest-passwordreset==1.1.0
django-silk==4.1.0
django-storages==1.11.1
django-tenant-schemas==1.10.0
django-tenant-users==0.3.12
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
drf-extra-fields==3.0.2
drf-yasg==1.20.0
gprof2dot==2021.2.21
gunicorn==20.0.4
honcho==1.0.1
idna==2.10
importlib-metadata==1.7.0
inclusive-django-range-fields==0.2.3
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.1
ordered-set==4.0.2
packaging==20.4
Pillow==8.0.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.3
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
sentry-sdk==0.17.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tenant-schemas-celery==1.0.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
vine==1.3.0
whitenoise==5.2.0
zipp==3.1.0
